Working on DB schemes

+++ config/ManageScheme.php
@@ -1,10 +1,14 @@
 <?php
 
+include_once(dirname(__DIR__)."/utils.php");
+include_once(__DIR__."/database.php");
+
 class ManageScheme
 {
     private $db;
-    private $dbName;
     private $create;
+    private $dbName;
+    const MAX_COLUMN_WIDTH = 27;
 
     public function __construct($info)
     {
@@ -18,7 +22,7 @@ class ManageScheme
     private function initDB($info)
     {
         try {
-            $this->db = new PDO('mysql:host=127.0.0.1;port=3306;dbname=Camagru', $info['username'], $info['password']);
+            $this->db = new PDO('mysql:host=mysql;port=3307;dbname=Matcha', $info['username'], $info['password']);
             $this->db->setAttribute(PDO::ATTR_EMULATE_PREPARES, true);
             $this->db->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
             $this->dbName = $info['db_name'];
@@ -28,6 +32,26 @@ class ManageScheme
         }
     }
 
+    private function isModelRegistered($table, $type)
+    {
+        if (!isset($this->sqlModel))
+            return (1);
+        foreach ($this->sqlModel as $model) {
+            if ($model->table == $table && $model->type == $type)
+                return (1);
+        }
+        return (0);
+    }
+
+    private function findModelQuery($table)
+    {
+        foreach ($this->sqlModel as $model) {
+            if ($model->table == $table && $model->type == "create")
+                return ($model->sql);
+        }
+        return (0);
+    }
+
     public function add($table, $type, $sql = null)
     {
       if (!isset($type) || $type == "" || !$this->isSupportedType($type))
@@ -35,13 +59,28 @@ class ManageScheme
       if (!isset($table) || $table == "")
         throw new Exception("Invalid table name !");
       if (($type == "reset" || $type == "delete") && $sql !== null)
-        throw new Exception("This type".$type."doesnt need an sql query !");
+        throw new Exception("This type".$type."doesnt need an sql query !". PHP_EOL);
       if (($type == "create" || $type == 'refresh') && $sql == null)
-        throw new Exception("This type : ".$type." need an sql query !");
+        throw new Exception("This type : ".$type." need an sql query !". PHP_EOL);
+      if (!isModelRegistered($table, $type))
+        throw new Exception($table."is already associated with a type !". PHP_EOL);
       $this->sqlModel[] = (object)['type' => $type, 'sql' => $sql, 'table' => $table];
     }
 
-    // reset ['table_name'] create['table_name'] delete['table_name'] ou create['*']
+    private function putError($table, $type, $msg)
+    {
+        $pad = self::MAX_COLUMN_WIDTH - strlen($table);
+        vprintf("%s %s".str_repeat(" ", $pad)."\e[31m%s\e[0m". PHP_EOL,
+        [$type, $table, $msg]);
+    }
+
+    private function putSuccess($table, $type, $msg)
+    {
+        $pad = self::MAX_COLUMN_WIDTH - strlen($table);
+        vprintf("%s %s".str_repeat(" ", $pad)."\e[92m%s\e[0m". PHP_EOL,
+        [$type, $table, 'sucess']);
+    }
+
     private function isSupportedType($type)
     {
         return ($type == 'delete' || $type == 'create' || $type == 'reset' || $type == 'refresh');
@@ -52,7 +91,6 @@ class ManageScheme
       if (!isset($this->sqlModel) || !is_array($this->sqlModel))
           throw new Exception("No sql model added !");
         foreach ($this->sqlModel as $model) {
-          $query = '';
           if ($model->type == "create"){
               $this->create($model->table, $model->sql);
           }
@@ -63,17 +101,24 @@ class ManageScheme
             $this->delete($model->table);
           }
           if ($model->type == "refresh"){
-            $this->delete($model->table);
-            $this->create($model->table);
+            if (!($query = findModelQuery($model->table)))
+                $this->putError($model->table, $model->type, "Error : ".$table." doesnt have a model query !");
+            else{
+                $this->delete($model->table);
+                $this->create($model->table, $query);
+            }
           }
         }
     }
 
-    private excecute($query)
+    private function execute($type, $table, $query)
     {
+      $pad = self::MAX_COLUMN_WIDTH - strlen($table);
       try {
         $this->db->exec($query);
+        $this->putSucess($table, $type, 'success');
       }catch (PDOException $e) {
+        $this->putError($table, $type, 'DB error');
         echo $e->getMessage();
         die();
       }
@@ -81,23 +126,56 @@ class ManageScheme
 
     private function create($table, $sql)
     {
+      $pad = self::MAX_COLUMN_WIDTH - strlen($table);
       $query.= "CREATE TABLE IF NOT EXISTS";
-      $query.= $table." ";
-      $query.= $sql;
-      $this->execute($query);
+      $query.= $table."(";
+      $query.= $sql.")";
+      $this->execute("Create", $table, $query);
     }
 
     private function reset($table)
     {
       $query.= "TRUNCATE TABLE ";
       $query.= $table;
-      $this->execute($query);
+      $this->execute("Reset", $table, $query);
     }
 
-    private function reset($table)
+    private function delete($table)
     {
       $query.= "DELETE TABLE IF EXISTS ";
       $query.= $table;
-      $this->execute($query);
+      $this->execute("Delete", $table, $query);
+    }
+
+    public function resetAll()
+    {
+        if (!isset($this->sqlModel) || !is_array($this->sqlModel))
+            throw new Exception("No sql model added !");
+        foreach ($this->sqlModel as $model) {
+            $this->reset($model->table);
+        }
+    }
+
+    public function deleteAll()
+    {
+        if (!isset($this->sqlModel) || !is_array($this->sqlModel))
+            throw new Exception("No sql model added !");
+        foreach ($this->sqlModel as $model) {
+                $this->reset($model->table);
+        }
+    }
+
+    public function refreshAll()
+    {
+        if (!isset($this->sqlModel) || !is_array($this->sqlModel))
+            throw new Exception("No sql model added !");
+        foreach ($this->sqlModel as $model) {
+            if (!($query = findModelQuery($model->table)))
+                $this->putError($model->table, $model->type, "Error : ".$table." doesnt have a model query !");
+            else{
+                $this->delete($model->table);
+                $this->create($model->table, $query);
+            }
+        }
     }
 }
